BUG: Use cache in _getintp_ctype, move tests.
